% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Atlas.R
\name{get_atlas}
\alias{get_atlas}
\title{get_atlas}
\usage{
get_atlas(import = TRUE)
}
\arguments{
\item{import}{TRUE if want to return data.frame}
}
\value{
list
}
\description{
get_atlas
}
\examples{
get_atlas(import=FALSE)
}
