% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Callejero.R
\name{get_callejero}
\alias{get_callejero}
\title{Get Callejero}
\usage{
get_callejero(year, ver)
}
\arguments{
\item{year}{string}

\item{ver}{string}
}
\value{
data.frame
}
\description{
Returns a data frame containing the complete spanish address catalogue.
For a given Address, the catalogue specifies among others:
Province Code, Municipality Code, District Code, Section Code, ZIP Code,
Street Name, Numeration...
}
\details{
For more information on the catalogue design, visit :
https://www.ine.es/ss/Satellite?c=Page&p=1259952026632&pagename=ProductosYServicios\%2FPYSLayout&cid=1259952026632&L=1
}
\examples{
get_callejero(year = "2018", ver = "01")
}
