% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cartography.R
\name{get_cartography}
\alias{get_cartography}
\title{get_cartography}
\usage{
get_cartography(year, import = TRUE)
}
\arguments{
\item{year}{string}

\item{import}{}
}
\value{
sf
}
\description{
get_cartography
}
