% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Atlas.R
\name{transform_atlas}
\alias{transform_atlas}
\title{Returns Atlas, an experimentaö data frame with Income Information from the spanish Tax Office}
\usage{
transform_atlas(object)
}
\arguments{
\item{object}{data.frame}
}
\value{
data.frame
}
\description{
Returns Atlas, an experimentaö data frame with Income Information from the spanish Tax Office
}
\examples{

}
