% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Callejero.R
\name{transform_callejero}
\alias{transform_callejero}
\title{transform_callejero}
\usage{
transform_callejero(object)
}
\arguments{
\item{object}{list containing data.frames tramero and vias}
}
\value{
data.frame
}
\description{
Takes list object containing tramero and vias and
returns joined dataset with the directory of spanish addresses
}
\examples{
transform_callejero(get_callejero("2018", "01"))
}
